Feature/notification ms patch (#55)
* notification patch

* notification patch tests

* correction

+++ notification-service/src/main/java/ba/unsa/etf/nwt/notificationservice/service/NotificationService.java
@@ -1,10 +1,15 @@
 package ba.unsa.etf.nwt.notificationservice.service;
 
 import ba.unsa.etf.nwt.notificationservice.dto.NotificationDTO;
+import ba.unsa.etf.nwt.notificationservice.exception.base.ForbiddenException;
+import ba.unsa.etf.nwt.notificationservice.exception.base.NotFoundException;
+import ba.unsa.etf.nwt.notificationservice.exception.base.UnprocessableEntityException;
 import ba.unsa.etf.nwt.notificationservice.model.Notification;
 import ba.unsa.etf.nwt.notificationservice.repository.NotificationRepository;
 import ba.unsa.etf.nwt.notificationservice.request.CreateNotificationRequest;
+import ba.unsa.etf.nwt.notificationservice.request.PatchNotificationRequest;
 import ba.unsa.etf.nwt.notificationservice.security.ResourceOwner;
+import ba.unsa.etf.nwt.notificationservice.utility.JsonNullableUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -28,15 +33,31 @@ public class NotificationService {
         return notification;
     }
 
-    public boolean existsById(UUID notificationId) {
-        return notificationRepository.existsById(notificationId);
-    }
-
     public void deleteById(UUID notificationId) {
+        if(!notificationRepository.existsById(notificationId))
+            throw new UnprocessableEntityException("Request body can not be processed. This notification doesn't exist");
+
         notificationRepository.deleteById(notificationId);
     }
 
     public Page<NotificationDTO> getNotificationsForUser(final UUID userId, final Pageable pageable) {
         return notificationRepository.findAllByUserId(userId, pageable);
+    }
+
+    public Notification patch(final Notification notification, final PatchNotificationRequest patchNotificationRequest) {
+        JsonNullableUtils.changeIfPresent(patchNotificationRequest.getRead(), notification::setRead);
+        return notificationRepository.save(notification);
+    }
+
+    public void checkUserId(UUID resourceOwnerId, UUID userId) {
+        if(!resourceOwnerId.equals(userId))
+            throw new ForbiddenException("Forbidden");
+    }
+
+    public Notification findById(UUID notificationId) {
+        if (notificationRepository.findById(notificationId).isPresent())
+            return notificationRepository.findById(notificationId).get();
+
+        throw new NotFoundException("Notification not found");
     }
 }
